Include the radial averaging too
